+++ sample/src/main/java/com/github/panpf/assemblyadapter/sample/ui/recycler/RecyclerGridDividerHorFragment.kt
@@ -220,6 +220,7 @@ class RecyclerGridDividerHorFragment : ToolbarFragment<FragmentRecyclerDividerHo
                 val size = dividerParams.dividerSize
                 val insets = Insets.allOf(dividerParams.dividerInsetsSize)
                 addAssemblyGridDividerItemDecoration {
+                    disableDefaultDivider()
                     if (dividerParams.isShowDivider) {
                         divider(Divider.colorRes(R.color.divider, size, insets)) {
                             personaliseByItemFactoryClass(