+++ sample/src/main/java/com/github/panpf/assemblyadapter/sample/ui/recycler/RecyclerLinearDividerVerFragment.kt
@@ -195,6 +195,7 @@ class RecyclerLinearDividerVerFragment :
                 val insets = Insets.allOf(dividerParams.dividerInsetsSize)
                 val dividerSize = dividerParams.dividerSize
                 addAssemblyLinearDividerItemDecoration {
+                    disableDefaultDivider()
                     if (dividerParams.isShowDivider) {
                         divider(Divider.colorRes(R.color.divider, dividerSize, insets)) {
                             personaliseByItemFactoryClass(