+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # new
 
 * fix: GridDividerItemDecoration and StaggeredGridDividerItemDecoration is divided evenly when the width of the item is the width of the parent container minus all dividers（`val itemSize=(parentWidth - (dividerSize * (spanCount+1))) / spanCount`），because the offset of the first item is too large, the width of the first item will be forced to become smaller
+* new: All DividerItemDecoration adds the `disableDefaultDivider(Boolean)` method, which can disable the function of using the system default divider when no divider is specified
 
 # v4.0.1